fix home page issue

+++ src/app/pages/home/flight-search-widget/flight-search-widget.component.ts
@@ -121,7 +121,8 @@ export class FlightSearchWidgetComponent implements OnInit {
     }
     this.route.queryParams.subscribe(params => {
 
-      if (Object.keys(params).length > 0) {
+      console.log("innn",params)
+      if (Object.keys(params).length > 0 && window.location.pathname=='/flight/search') {
         //delete BehaviorSubject in the listing page
         this.homeService.removeToString();
